requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
